+++ test/tasks/ods-build-typescript_test.go
@@ -40,6 +40,7 @@ func TestTaskODSBuildTypescript(t *testing.T) {
 						filepath.Join(pipelinectxt.SonarAnalysisPath, "analysis-report.md"),
 						filepath.Join(pipelinectxt.SonarAnalysisPath, "issues-report.csv"),
 						filepath.Join(pipelinectxt.LintReportsPath, "report.txt"),
+						"docker/dist/src/index.js",
 					)
 
 					wantLogMsg := "No sonar-project.properties present, using default:"
@@ -84,7 +85,6 @@ func TestTaskODSBuildTypescript(t *testing.T) {
 				},
 			},
 			"fail linting typescript app and generate lint report": {
-				Timeout:             20 * time.Minute,
 				WorkspaceDirMapping: map[string]string{"source": "typescript-sample-app-lint-error"},
 				PreRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
 					wsDir := ctxt.Workspaces["source"]
@@ -112,6 +112,58 @@ func TestTaskODSBuildTypescript(t *testing.T) {
 					}
 				},
 				WantSetupFail: true,
+			},
+			"build backend typescript app": {
+				Timeout:             20 * time.Minute,
+				WorkspaceDirMapping: map[string]string{"source": "typescript-sample-app"},
+				PreRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+					ctxt.ODS = tasktesting.SetupGitRepo(t, ctxt.Namespace, wsDir)
+					ctxt.Params = map[string]string{
+						"copy-node-modules": "true",
+					}
+				},
+				WantRunSuccess: true,
+				PostRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+
+					checkFilesExist(t, wsDir,
+						filepath.Join(pipelinectxt.XUnitReportsPath, "report.xml"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "clover.xml"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "coverage-final.json"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "lcov.info"),
+						filepath.Join(pipelinectxt.SonarAnalysisPath, "analysis-report.md"),
+						filepath.Join(pipelinectxt.SonarAnalysisPath, "issues-report.csv"),
+						filepath.Join(pipelinectxt.LintReportsPath, "report.txt"),
+						"docker/dist/node_modules/",
+					)
+				},
+			},
+			"build typescript app with custom build directory": {
+				Timeout:             20 * time.Minute,
+				WorkspaceDirMapping: map[string]string{"source": "typescript-sample-app-build-dir"},
+				PreRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+					ctxt.ODS = tasktesting.SetupGitRepo(t, ctxt.Namespace, wsDir)
+					ctxt.Params = map[string]string{
+						"build-dir": "build",
+					}
+				},
+				WantRunSuccess: true,
+				PostRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+
+					checkFilesExist(t, wsDir,
+						filepath.Join(pipelinectxt.XUnitReportsPath, "report.xml"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "clover.xml"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "coverage-final.json"),
+						filepath.Join(pipelinectxt.CodeCoveragesPath, "lcov.info"),
+						filepath.Join(pipelinectxt.SonarAnalysisPath, "analysis-report.md"),
+						filepath.Join(pipelinectxt.SonarAnalysisPath, "issues-report.csv"),
+						filepath.Join(pipelinectxt.LintReportsPath, "report.txt"),
+						"docker/dist/src/index.js",
+					)
+				},
 			},
 		})
 }